+++ src/plone/app/mosaic/browser/static/js/mosaic.layout.js
@@ -435,83 +435,6 @@ define([
             $(this).mosaicSelectTile();
         });
 
-        $($.mosaic.document).on("click", ".mosaic-close-icon", function () {
-
-            // Get tile config
-            var tile_config = $(this).parents(".mosaic-tile").mosaicGetTileConfig();
-
-            // Check if app tile
-            if (tile_config.tile_type === 'app') {
-
-                // Get url
-                var tile_url = $(this).parents(".mosaic-tile").find('.tileUrl').html();
-
-                // Remove tags
-                $.mosaic.removeHeadTags(tile_url);
-
-                // Calc delete url
-                var url = tile_url.split('?')[0];
-                url = url.split('@@');
-                var tile_type_id = url[1].split('/');
-                url = url[0] + '@@delete-tile/' + tile_type_id[0] + '/' + tile_type_id[1];
-                // Calc absolute delete url
-                if (url.match(/^\.\/.*/)) {
-                    url = $.mosaic.options.context_url + url.replace(/^\./, '');
-                }
-
-                // Ajax call to remove tile
-                $.ajax({
-                    type: "GET",
-                    url: url,
-                    success: function (value) {
-                        var authenticator = $(value).find('[name="_authenticator"]').val();
-                        $.ajax({
-                            type: "POST",
-                            url: url,
-                            data: {
-                                'buttons.delete': 'Delete',
-                                '_authenticator': authenticator
-                            },
-                            success: function(value) {
-                                /*
-                                 $.plone.notify({
-                                     title: "Info",
-                                     message: "Application tile removed",
-                                     sticky: false
-                                 });
-                                */
-                            }
-                        });
-                    }
-                });
-            }
-
-            // Remove empty rows
-            $.mosaic.options.panels.find(".mosaic-empty-row").remove();
-
-            // Get original row
-            var original_row = $(this).parents(".mosaic-tile").parent().parent();
-
-            // Save tile value
-            $.mosaic.saveTileValueToForm(tile_config.name, tile_config);
-
-            // Remove current tile
-            $(this).parent().remove();
-
-            $.mosaic.undo.snapshot();
-
-            // Cleanup original row
-            original_row.mosaicCleanupRow();
-
-            // Add empty rows
-            $.mosaic.options.panels.mosaicAddEmptyRows();
-
-            // Set toolbar
-            $.mosaic.options.toolbar.trigger("selectedtilechange");
-            $.mosaic.options.toolbar.mosaicSetResizeHandleLocation();
-        });
-
-
         // On click open overlay
         $($.mosaic.document).on("click", ".mosaic-info-icon", function () {
 
@@ -696,13 +619,6 @@ define([
                     $($.mosaic.document.createElement("div"))
                         .addClass("mosaic-tile-control mosaic-drag-handle")
                 );
-            }
-
-            // If tile is removable
-            if ($(this).hasClass("removable") && $.mosaic.options.can_change_layout) {
-
-                // Add close icon
-                $(this).prepend('<div class="mosaic-tile-control mosaic-close-icon"></div>');
             }
 
             // Add settings icon